// src/model/createdCell.ts
/**
 * Generated by orval v7.5.0 🍺
 * Do not edit manually.
 * Key Station Documentation
 * API станции для выдачи ключей по пропуску преподавателей
 * OpenAPI spec version: 1.0
 */
import type { CreatedCustomerDto } from './createdCustomerDto';

export interface CreatedCell {
  id: string;
  floor: number;
  classroom: number;
  isLocked: boolean;
  createdAt: string;
  updatedAt: string;
  customerId: string;
  customer: CreatedCustomerDto;
}